### 2024-03-25 #6535 @spholmes
Hi Olivia,
I just volunteered for a different paper and this is very far from my focus,
so I'll pass on this one.
Best
Susan

On Mon, Mar 25, 2024 at 9:24 AM Olivia Guest ***@***.***>
wrote:

> @emdupre <https://github.com/emdupre> @spholmes
> <https://github.com/spholmes> what are your capacities looking like for
> this? If you can take it please do, if not let me know. ☺️
>
