### 2024-03-25 70bc3f2
Attempt to fix paper heading for JOSS parsing.
