### 2024-04-01 #6535 @marcosvital
Hi, @florisvanvugt , I will be the editor of your submission.

Please take a look at the manuscript proof you generated above, and check if it seems ok and ready to review. If you make any changes or corrections on it you can create a new version anytime.

At the same time, we can start assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet). Also, you can search potential reviewers [here](https://reviewers.joss.theoj.org/lookup) - or just let me know if you prefer for me to do it, and I'll look into it.
