+++ paper/paper.bib
@@ -70,11 +70,25 @@ Publisher: Nature Publishing Group},
 }
 
 
-@online{opensignals,
-	title = {{OpenSignals} {\textbar} Data Visualization Software {\textbar} Bitalino},
-	url = {https://bitalino.com/en/software},
-	urldate = {2020-03-12},
+@inproceedings{opensignals,
+author="Braga, Lucas C. M. F.
+and Castro, Maria Claudia F.
+and Avelino, Valter F.",
+editor="Costa-Felix, Rodrigo
+and Machado, Jo{\~a}o Carlos
+and Alvarenga, Andr{\'e} Victor",
+title="Educational Platform for Physiological Signal Measurements",
+booktitle="XXVI Brazilian Congress on Biomedical Engineering",
+year="2019",
+publisher="Springer Singapore",
+address="Singapore",
+pages="671--677",
+abstract="The constant evolution in the field of Biomedical Engineering requires betterment resources of bioelectrical signal analysis for Biomedical Engineering student demonstration. This project aims to develop a didactic platform intended to bioelectrical signals acquisition and processing. To perform this task, a specific hardware was used to the acquisition of bioelectrical signals whose characteristics are their versatility, be an open-source and be a low-cost device. Among the available options, it was selected the hardware called BITalino. To verify the efficiency of this hardware a characterization process was carried out, which consisted of analyzing the accuracy, frequency response and noise level associated with the acquisition of three types of bioelectrical signals (EMG, ECG, EEG). This work developed an interface for acquisition and conditioning of bioelectrical signals, based on LabVIEW (National Instruments®) software environment. The interface development process was divided into 3 parts: communication interface with the device; data conversion process and definition of signal conditioning processes. The project{\textasciiacute}s focus was accomplished with the development of a didactic platform, which the student can add noise to the signal, choose filters and perform frequency analysis, helping them to assimilate all the signal conditioning process.",
+isbn="978-981-13-2517-5"
 }
+
+
+
 
 
 @article{Cox_1996,
