+++ paper/paper.md
@@ -51,7 +51,7 @@ Human biosignals such as breathing, cardiac rhythms or skin conductance contain 
 
 There is increasing interest on the part of the neuroscience and psychology research community in biosignals, that is, measurements of cardiac activity (typically from the electrocardiogram, ECG), electrodermal activity (EDA), respiration, and others [@Massaro2016;@Varga2017;@Horvers2021;@PosadaQuintero2020]. There are now wonderful packages for biosignal preprocessing (e.g. neurokit [@neurokit]) and analyzing (e.g. biopeaks [@Brammer2020]). However, progress is hampered by the proliferation of a multitude of file formats (EDF, XDF, OpenSignals, BDF, CSV, Acknowledge ACQ, etc.). Existing software packages typically read only one or two of these formats, requiring researchers to convert between formats which is tedious and error-prone, or simply impossible when using read-only libraries. Furthermore, data in these different formats is typically organized differently, requiring researchers to reorganize their code to cater to different formats. 
 
-Individual Python packages exist that can each read single data formats (e.g. [pyxdf](https://pypi.org/project/pyxdf/) or [pyedflib](https://pyedflib.readthedocs.io/en/latest/)). However, each makes the data available in a different structure. This means that pipelines have to be changed when switching from one data format to another, which is tedious and error-prone. The situation is further complicated by the fact that different file formats make different assumptions about the data structure: in some formats, multiple signals in a file are forced to have the same sampling rate (e.g. OpenSignals) whereas in other formats sampling rates can vary (e.g. XDF). In some cases the signals are supposed to have the same onset time (e.g. EDF) whereas other formats allow different onset times requiring re-aligning (e.g. XDF). All this makes conversion cumbersome and errors can easily slip in. 
+Individual Python packages exist that can each read single data formats (e.g. [pyxdf](https://pypi.org/project/pyxdf/) or [pyedflib](https://pyedflib.readthedocs.io/en/latest/)). However, each makes the data available in a different structure. This means that pipelines have to be changed when switching from one data format to another, which is tedious and error-prone. The situation is further complicated by the fact that different file formats make different assumptions about the data structure: in some formats, multiple signals in a file are forced to have the same sampling rate (e.g. OpenSignals [@opensignals]) whereas in other formats sampling rates can vary (e.g. XDF). In some cases the signals are supposed to have the same onset time (e.g. EDF) whereas other formats allow different onset times requiring re-aligning (e.g. XDF). All this makes conversion cumbersome and errors can easily slip in. 
 
 This state of affairs also hampers the development of unified, reproducible pipelines that can be shared between research groups across the globe. Increasingly, the field calls for sharing of data analysis pipelines between research groups as an indispensible step to much-needed reproducibility [@Wratten2021]. In addition, sharing analysis pipelines rather than reinventing the wheel allows for more efficient use of scientists' time.
 
