### 2024-04-11 #6535 @marcosvital
> Thanks for the invitation! I believe I would have a conflict of interest, as the submitting author and I have been collaborating.

Thank you of letting us know, @kinleyid. Recent academic collaboration in funded research or published work is indeed considered a potential conflict of interest [following JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
