### 2024-04-11 e9ac132
Creating a simple HTML report generator.
