### 2024-04-13 f747a9f
Tweaking the HTML report style.
