### 2024-04-20 3d4c53c
Minor tweak that makes the HTML file easier to read because full and small previews have different sizes.
