### 2024-04-24 0d64136
Adding list of supported file types to the open file dialog.
