### 2024-05-04 e5c94e9
Fixed bug reading invalid last row in B3 data.
