### 2024-05-17 c307517
Updated LSL reading script to correct minor bug.
