+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "biobabel"
-version = "1.0.3"
+version = "1.0.4"
 authors = [
   { name="Floris van Vugt", email="ftvanvugt@gmail.com" },
 ]
