### 2024-05-30 #6535 @danielmk
Dear Marcos, I've agreed to review another JOSS submission and I don't
think I have bandwidth to handle two at the same time so I have to decline
to review biobabel.

Best,
Daniel

On Tue, May 28, 2024, 22:40 Marcos Vital (LEQ-UFAL) <
***@***.***> wrote:

> Dear @hooman650 <https://github.com/hooman650>, @JanCBrammer
> <https://github.com/JanCBrammer> and @danielmk
> <https://github.com/danielmk>, would you be willing to review this
> submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and once there are enough reviewers we
> will move on to a new issue where the review will take place. If you are
> not able to review, please let us know if you have recommendations for
> other reviewers that could help with this submission.
>
