### 2024-06-04 b6c6f5b
Declaring version number dynamically.
