+++ src/biobabel/__init__.py
@@ -1,3 +1,6 @@
+__version__ = '1.0.4'
+
+
 import biobabel.io
 from biobabel.io import load
 
