### 2024-06-13 #6884 @editorialbot
[REVIEW]: biobabel: a unified interface for reading a plethora of file formats for biosignals such as cardiac, respiration, electrodermal data
**Submitting author:** @florisvanvugt (<a href="http://orcid.org/0000-0003-1739-0175">Floris Tijmen van Vugt</a>)
**Repository:** https://github.com/florisvanvugt/biobabel
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
