### 2024-06-13 #6884 @marcosvital
Dear @sappelhoff, @finsberg and @ntolley, thank you again for accepting review this submission for JOSS.

The reviewing process is checklist based, and instructions were already posted above by the editorial bot - but let me know if you need any assistance, ok?

Also, you can tag @florisvanvugt if you have specific questions about the manuscript.
