### 2024-06-14 #6884 @ntolley
@florisvanvugt going over the software and everything is looking clear and documented!

A major piece that is missing right now is an appropriate testing suite. Please see the guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html#tests

While I do indeed see a test folder, this is more of a tutorial/example that only covers the functionality for a single file type. Given that the point of this software is to integrate many disparate file types, at a minimum I think the tests should ensure all of the loading functions work for each supported file types.

My major recommendation would be to:
1) Generate or add  _very_ small files (size is important since once they are committed it will be carries in the git history) for each supported datatype
2) Set up `.py` testing files with formatted to work with pytest: https://docs.pytest.org/en/7.1.x/getting-started.html
3) Loop through these files and ensure that they can be loaded without error.

In the end you should be able to run `pytest test_loading.py` or something similar and quickly verify that your code is functioning correctly.
