### 2024-06-14 #6884 @ntolley
If you see yourself building out this software further, it is rather easier to set up automated testing through github workflows: https://docs.github.com/en/actions/automating-builds-and-tests/about-continuous-integration

While I think this is outside the scope of this JOSS submission, I just wanted to point you in the right direction if you hadn't seen this before
