### 2024-06-14 florisvanvugt/biobabel#2 @finsberg
[JOSS REVIEW] Add Contributing Guide
As listed in the [Review Checklist](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) there should be a Contributing Guide. This is typically done by adding a `CONTRIBUTING.md` (see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors for more info)

This is related to https://github.com/openjournals/joss-reviews/issues/6884
