### 2024-06-14 florisvanvugt/biobabel#3 @finsberg
[JOSS REVIEW] Add automated tests
This is also mentioned in https://github.com/openjournals/joss-reviews/issues/6884#issuecomment-2167113578 by @ntolley, which I think did an excellent job explaining what is needed. I will just make it an issue here for reference.

You have a folder called `tests` but this looks more like examples of how to use the software. You should write some tests than ensures that the program is working as expected. You should also set up a CI (using e.g GitHub actions) to run the tests every time  you push to the repo.

This is related to https://github.com/openjournals/joss-reviews/issues/6884
