### 2024-06-14 florisvanvugt/biobabel#4 @finsberg
[JOSS REVIEW] Missing dependency `neurokit2`
I tried to install and run the examples in the `tests` folder, and it looks like I am missing a dependency called `neurokit2`. If this is a  dependency that is only required to run the examples you could for examples add it as an optional dependency in the `pyproject.toml` (similar to what you have done with `pybdf`).

This is related to https://github.com/openjournals/joss-reviews/issues/6884#issuecomment-2167113578
