### 2024-06-14 florisvanvugt/biobabel#5 @finsberg
[JOSS REVIEW] Add help to commands
I think it would be nice if you could add a help flag to the commands that you have implemented. Rather than using `sys.argv` to parse the arguments from the command line you could use [`argparse`](https://docs.python.org/3/library/argparse.html) instead which is really the right tool for this job. For `bioinfo = "biobabel:info"` for example, you could instead do
```python
import argparse

def get_file():
    """
    Figure out which file wants to be opened.
    First check the command line if a file was indicated.
    If not, show a file selection dialog window to allow
    the user to select one.
    Open the file, and then print a summary.
    """
    parser = argparse.ArgumentParser(description="Display information about a file.")
    parser.add_argument("filename", nargs="?", help="The file to open")
    args = parser.parse_args()

    if args.filename:
        fname = sys.argv[1]
    else:
        fname = ask_bio_file()

    if not fname:
        print("You need to select a file. Exiting now.")
        sys.exit(-1)

    if not os.path.exists(fname):
        print("File {} does not seem to exist. Exiting now.".format(fname))
        sys.exit(-1)

    bio = biobabel.io.load(fname)
    bio.print()
    bio.meta["filename"] = fname
    return bio
```
Note that that this is not a requirement but rather a recommendation.

This is related to https://github.com/openjournals/joss-reviews/issues/6884
