+++ docs/source/api.rst
@@ -1,0 +1,7 @@
+API
+===
+
+.. autosummary::
+   :toctree: generated
+
+   biobabel
