+++ docs/source/index.rst
@@ -1,10 +1,9 @@
-Welcome to Lumache's documentation!
+Welcome to Biobabel's documentation
 ===================================
 
-**Lumache** (/lu'make/) is a Python library for cooks and food lovers
-that creates recipes mixing random ingredients.
-It pulls data from the `Open Food Facts database <https://world.openfoodfacts.org/>`_
-and offers a *simple* and *intuitive* API.
+**Biobabel** is a Python library providing a unified interface for reading a plethora of physiology file formats.
+
+Many different file formats exist for physiology signals such as cardiac (ECG, PPG) or respiratory data. Many packages exist in Python to read these formats but they load into different data structures. This package is a wrapper of sorts for already existing tools providing a unified easy-to-use interface.
 
 Check out the :doc:`usage` section for further information, including
 how to :ref:`installation` the project.
