### 2024-06-20 48e865c
Changing usage and hopefully importing the package itself for doc generation.
