+++ .readthedocs.yaml
@@ -8,6 +8,10 @@ build:
 python:
   install:
     - requirements: docs/requirements.txt
+    # Install our python package before building the docs
+    - method: pip
+      path: .
 
 sphinx:
   configuration: docs/source/conf.py
+  fail_on_warning: true
