### 2024-06-20 cd0eb6d
Adding numpy requirement for docs.
