+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 sphinx==7.1.2
 sphinx-rtd-theme==1.3.0rc1
+numpy==1.26.1
