+++ docs/source/conf.py
@@ -1,5 +1,11 @@
 # Configuration file for the Sphinx documentation builder.
 
+# -- Import relevant source
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../..'))
+
+
 # -- Project information
 
 project = 'Biobabel'
