+++ docs/source/index.rst
@@ -18,4 +18,5 @@ Contents
 .. toctree::
 
    usage
+   biodata
    api
