+++ docs/source/biodata.rst
@@ -1,0 +1,7 @@
+.. _biodata:
+
+Biodata
+=======
+
+.. autoclass:: biobabel.Biodata
+    :members:
