+++ src/biobabel/biodata.py
@@ -46,12 +46,16 @@ class Biodata:
         """ 
         Add metadata.
 
-        k : the metadata key 
-        v : the metadata value
-        replace : if true, replace any existing metadata key value, otherwise add a new key
+        :param k: str, the metadata key 
+        :param v: str, the metadata value
+        :param replace: bool, if `True`, replace any existing metadata key value, otherwise add a new key
 
         Example:
-        biodata.add_meta('date','2024-03-22')
+
+        .. code-block:: python
+
+           biodata.add_meta('date','2024-03-22')
+
         """
         kk = k
         if not replace:
@@ -64,7 +68,9 @@ class Biodata:
         Return a channel id that is unique (does not exist yet), 
         as close as possible to the given id.
 
-        ident : the id to start from
+        :param ident: str, the id to start from
+        :return: a channel id that does not exist yet
+        :rtype: str
 
         If ident is already unique, it will be returned.
         """
@@ -81,10 +87,10 @@ class Biodata:
     
     def add_channel(self,hdrdat):
         """ 
-        Add the channel with specified header information and data.
+        Create a new channel with specified header information and data.
 
-        hdrdat : is a tuple (hdr,dat) where hdr is a dict containing the
-        header and dat is the data stream itself (a one-dimensional array).
+        :param hdrdat: is a tuple (hdr,dat) where hdr is a dict containing the header and dat is the data stream itself (a one-dimensional array).
+
         """
         hdr,dat   = hdrdat
         newid     = hdr['id'].replace('/','_') # slashes don't work because HDF5 gets confused
@@ -93,14 +99,38 @@ class Biodata:
 
 
     def find(self,crit={}):
-        """ Find all channels that satisfy the metadata criteria and return them as (hdr,dat) """
+        """ 
+        Find all channels that satisfy the metadata criteria and return them as (hdr,dat) 
+
+        :param crit: a dictionary containing key-value pairs for searching the channels.
+        :returns: the channels (hdr,dat) satisfying the search criteria.
+
+        Example:
+
+        .. code-block:: python
+
+           biodata.find({'modality':'ecg'}) # will return all channels where modality equals ecg
+
+        """
         return [ self.get(ch) for ch in self.find_channels(crit) ]
             
             
         
         
     def find_channels(self,crit={}):
-        """ Find the channel id's satisfying certain criteria """
+        """ 
+        Find the channel id's satisfying certain criteria.
+
+        :param crit: a dictionary containing key-value pairs for searching the channels.
+        :returns: the channel IDs satisfying the search criteria.
+        :rtype: str list
+
+        Example:
+
+        .. code-block:: python
+
+           biodata.find_channels({'modality':'ecg'}) # will return all channel IDs for channels whose modality equals ecg
+        """
         chans = []
         for hdr,_ in self.channels:
             ok = True
@@ -113,7 +143,12 @@ class Biodata:
 
 
     def get_participants(self):
-        """ Return a list of participants in the current object """
+        """ 
+        Return a list of participants in the current object 
+
+        :returns: a list of participant IDs
+        :rtype: str list
+        """
         part = []
         for hdr,_ in self.channels:
             p= hdr.get('participant',None)
@@ -123,7 +158,12 @@ class Biodata:
         return part
 
     def get(self,chid):
-        """ Get a particular channel data """
+        """ 
+        Get a particular channel data 
+        
+        :param chid: str, the channel ID
+        :returns: (hdr,dat) tuple containing the header and data for the given channel, respectively.
+        """
         for hdr,dat in self.channels:
             if hdr.get('id',None)==chid:
                 return hdr,dat
@@ -131,13 +171,27 @@ class Biodata:
         return None,None
 
     def get_time(self,chid):
[…]
