+++ docs/source/index.rst
@@ -18,5 +18,6 @@ Contents
 .. toctree::
 
    usage
+   cli
    biodata
    api
