### 2024-06-20 5999a5f
Adding screenshots and more details about command line usage.
