+++ docs/source/index.rst
@@ -5,12 +5,8 @@ Welcome to Biobabel's documentation
 
 Many different file formats exist for physiology signals such as cardiac (ECG, PPG) or respiratory data. Many packages exist in Python to read these formats but they load into different data structures. This package is a wrapper of sorts for already existing tools providing a unified easy-to-use interface.
 
-Check out the :doc:`usage` section for further information, including
-how to :ref:`installation` the project.
+Check out the :doc:`usage` section for basic usage information, and :doc:`cli` for usage from the console (terminal). 
 
-.. note::
-
-   This project is under active development.
 
 Contents
 --------
