### 2024-07-08 florisvanvugt/biobabel#6 @sappelhoff
I agree with everything @finsberg says. I would furthermore recommend you adopt a particular style for docstrings (and use it consistently), and that you add [module docstrings](https://stackoverflow.com/q/2557110/5201771) to each `.py` file. I recommend [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html).
