### 2024-07-08 #6884 @sappelhoff
I think @ntolley and @finsberg have done a great job at their reviews and I agree with all the points they have raised so far.

I think this package looks promising and helpful, and for the review to proceed (tick more boxes from my checklist), I would need to see an automatic test suite, as discussed here:

- https://github.com/openjournals/joss-reviews/issues/6884#issuecomment-2167110653
- https://github.com/florisvanvugt/biobabel/issues/3
