### 2024-07-08 florisvanvugt/biobabel#8 @sappelhoff
[JOSS REVIEW] Pull functions out of  __init__.py
- xref: https://github.com/openjournals/joss-reviews/issues/6884

You have defined several functions in your `__init__.py` file, for example: https://github.com/florisvanvugt/biobabel/blob/76efa3fcf8b545e2dc971b432573c979f98ad56e/src/biobabel/__init__.py#L16-L21

This is the first time that I have seen this practice. As far as I know, `__init__.py` files are usually either empty, or contain "import statements", or variable definitions such as `__version__`. I would recommend to follow the "common practice" and define functions in dedicated modules (.py files), to then import them and expose them at the wanted "module level" via `__init__.py`.

For example, have a look at this: https://github.com/mne-tools/mne-bids/blob/754efa9fa35cfd293faf918860e325a59bd7e0f0/mne_bids/__init__.py#L11

The line in `__init__.py` says:

```python
from mne_bids.report import make_report
```

which means that users will be able to call `mne_bids.make_report` directly, instead of going via `mne_bids.report.make_report`.
