### 2024-07-08 florisvanvugt/biobabel#9 @sappelhoff
[JOSS REVIEW] fix API section in documentation
- xref: https://github.com/openjournals/joss-reviews/issues/6884

The documentation pages contain two sections for API docs:

1. `biobabel`
2. `API`

Ideally, `biobabel` should just be listed under `API`.

Furthermore, under `API` there are currently some functions that are not fully documented, see this screenshot as an example:
