### 2024-07-30 #6884 @finsberg
@marcosvital There has been no response from the author on any of my issues, which was opened six weeks ago. I think it should be expected that the response time for an open source project should be faster than six week, and therefore I would like to reject this submission for publication in JOSS.
