+++ src/biobabel/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.5'
+__version__ = '1.0.6'
 
 
 import biobabel.io
