### 2024-08-29 #6884 @marcosvital
Dear @finsberg and @sappelhoff, I'm sorry about the delayed reply.

We don't have a specific response time that should lead to automatic rejection, but I understand that we should expect some response from the author in a timely fashion or an expected timeline to adress the issues. Since the author did not answer here, I'll try to reach him elsewhere and will let you know what might happen next.

I'm sorry about this, and would like to thank both of you and @ntolley for the effort put so far into this review.
