### 2024-09-25 249bf94
Numpy syntax update np.inf
