### 2024-11-12 29e3780
Adding contributing guidelines.
