### 2024-11-12 florisvanvugt/biobabel#2 @florisvanvugt
Thank you for the suggestion. A contributing guide is now included. https://github.com/florisvanvugt/biobabel/commit/29e37805dbdcdf244a90d76d2763c068d5152bd4
