### 2024-11-12 78a788c
Marking neurokit2 as optional dependency for example usage.
