### 2024-11-12 florisvanvugt/biobabel#4 @florisvanvugt
Thank you for spotting that. Indeed, `neurokit2` is not required for the functionality of the package, but it is referred to for illustration purposes in the example usage script. I have followed your suggestion and marked it as an optional dependency.
https://github.com/florisvanvugt/biobabel/commit/78a788c3d77402440f12f527235fc2d6513a3a97
