### 2024-11-12 ddef9de
Adding command line argument parsing.
