### 2024-11-12 florisvanvugt/biobabel#5 @florisvanvugt
I agree this is a great idea. It really helps when you're running the script on the command line and you don't have the documentation at hand. Fixed in: https://github.com/florisvanvugt/biobabel/commit/ddef9de936d5847a52a1c5e5e1cd1005cfe8ea13
